Style: 하단 메뉴 분리
